renames wrapHits
