+++ x-pack/plugins/security_solution/server/lib/detection_engine/signals/search_after_bulk_create.ts
@@ -7,9 +7,7 @@
 
 import { identity } from 'lodash';
 import { SortResults } from '@elastic/elasticsearch/api/types';
-import { Logger } from '@kbn/logging';
 import { singleSearchAfter } from './single_search_after';
-import { filterDuplicateRules, filterDuplicateSignals } from './single_bulk_create';
 import { filterEventsAgainstList } from './filters/filter_events_against_list';
 import { sendAlertTelemetryEvents } from './send_telemetry_events';
 import {
@@ -20,17 +18,8 @@ import {
   mergeReturns,
   mergeSearchResults,
   getSafeSortIds,
-  generateId,
 } from './utils';
-import {
-  SearchAfterAndBulkCreateParams,
-  SearchAfterAndBulkCreateReturnType,
-  SignalSearchResponse,
-  WrapHits,
-  WrappedSignalHit,
-} from './types';
-import { buildBulkBody } from './build_bulk_body';
-import { BuildRuleMessage } from './rule_messages';
+import { SearchAfterAndBulkCreateParams, SearchAfterAndBulkCreateReturnType } from './types';
 
 // search_after through documents and re-index using bulk endpoint.
 export const searchAfterAndBulkCreate = async ({
@@ -41,15 +30,13 @@ export const searchAfterAndBulkCreate = async ({
   listClient,
   logger,
   eventsTelemetry,
-  id,
   inputIndexPattern,
-  signalsIndex,
   filter,
   pageSize,
   buildRuleMessage,
   enrichment = identity,
   bulkCreate,
-  wrapSignals,
+  wrapHits,
 }: SearchAfterAndBulkCreateParams): Promise<SearchAfterAndBulkCreateReturnType> => {
   const ruleParams = ruleSO.attributes.params;
   let toReturn = createSearchAfterReturnType();
@@ -160,7 +147,7 @@ export const searchAfterAndBulkCreate = async ({
             );
           }
           const enrichedEvents = await enrichment(filteredEvents);
-          const wrappedDocs = wrapSignals(enrichedEvents.hits.hits);
+          const wrappedDocs = wrapHits(enrichedEvents.hits.hits);
 
           const {
             bulkCreateDuration: bulkDuration,
@@ -208,60 +195,4 @@ export const searchAfterAndBulkCreate = async ({
   logger.debug(buildRuleMessage(`[+] completed bulk index of ${toReturn.createdSignalsCount}`));
   toReturn.totalToFromTuples = tuplesToBeLogged;
   return toReturn;
-};
-
-export const buildWrappedSignalsFactory = ({
-  ruleSO,
-  signalsIndex,
-}: {
-  ruleSO: SearchAfterAndBulkCreateParams['ruleSO'];
-  signalsIndex: string;
-}): WrapHits => (events) => {
-  const wrappedDocs: WrappedSignalHit[] = events.flatMap((doc) => [
-    {
-      _index: signalsIndex,
-      _id: generateId(
-        doc._index,
-        doc._id,
-        doc._version ? doc._version.toString() : '',
-        ruleSO.attributes.params.ruleId ?? ''
-      ),
-      _source: buildBulkBody(ruleSO, doc),
-    },
-  ]);
-
-  return filterDuplicateSignals(ruleSO.id, wrappedDocs);
-};
-
-export const filterAndWrapDocuments = ({
-  buildRuleMessage,
-  enrichedEvents,
-  id,
-  logger,
-  ruleSO,
-  signalsIndex,
-}: {
-  buildRuleMessage: BuildRuleMessage;
-  enrichedEvents: SignalSearchResponse;
-  id: string;
-  logger: Logger;
-  ruleSO: SearchAfterAndBulkCreateParams['ruleSO'];
-  signalsIndex: string;
-}) => {
-  enrichedEvents.hits.hits = filterDuplicateRules(id, enrichedEvents);
-  logger.debug(buildRuleMessage(`about to bulk create ${enrichedEvents.hits.hits.length} events`));
-
-  const wrappedDocs: WrappedSignalHit[] = enrichedEvents.hits.hits.flatMap((doc) => [
-    {
-      _index: signalsIndex,
-      _id: generateId(
-        doc._index,
-        doc._id,
-        doc._version ? doc._version.toString() : '',
-        ruleSO.attributes.params.ruleId ?? ''
-      ),
-      _source: buildBulkBody(ruleSO, doc),
-    },
-  ]);
-  return wrappedDocs;
 };
